add AdultCensus, BikeRegression and tests to benchmark games.
