+++ shapiq/games/__init__.py
@@ -1,10 +1,12 @@
 """This module contains sample game functions for the shapiq package."""
 
 from .base import Game
+from .benchmark_tabular import AdultCensus, BikeRegression, CaliforniaHousing
 from .dummy import DummyGame
+from .image_classifier import ImageClassifierGame
 from .imputer import MarginalImputer
 from .sentiment_language import SentimentClassificationGame
-from .tabular import CaliforniaHousing, FeatureSelectionGame, LocalExplanation
+from .tabular import FeatureSelectionGame, LocalExplanation
 
 __all__ = [
     "DummyGame",
@@ -14,4 +16,7 @@ __all__ = [
     "LocalExplanation",
     "FeatureSelectionGame",
     "CaliforniaHousing",
+    "BikeRegression",
+    "AdultCensus",
+    "ImageClassifierGame",
 ]
